video-ads-generation/requirements.txt
requests
python-dotenv
pandas
openpyxl
streamlit>=1.28.0
